+++ src/com/intellij/idea/plugin/hybris/impex/psi/references/ImpexFunctionTSAttributeReference.kt
@@ -1,10 +1,11 @@
 /*
- * This file is part of "hybris integration" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as 
- * published by the Free Software Foundation, either version 3 of the 
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
  * License, or (at your option) any later version.
  *
  * This program is distributed in the hope that it will be useful,
@@ -18,7 +19,7 @@
 
 package com.intellij.idea.plugin.hybris.impex.psi.references
 
-import com.intellij.codeInsight.lookup.LookupElementBuilder
+import com.intellij.codeInsight.lookup.LookupElement
 import com.intellij.idea.plugin.hybris.common.HybrisConstants
 import com.intellij.idea.plugin.hybris.impex.psi.ImpexFullHeaderParameter
 import com.intellij.idea.plugin.hybris.impex.psi.ImpexParameter
@@ -49,7 +50,7 @@ class ImpexFunctionTSAttributeReference(owner: ImpexParameter) : TSReferenceBase
         return TextRange.from(element.text.indexOf(attributeName), attributeName.length)
     }
 
-    override fun getVariants(): Array<LookupElementBuilder> {
+    override fun getVariants(): Array<LookupElement> {
         // if inline type already present we should not suggest any other types
         if (element.inlineTypeName != null) return emptyArray()
         return TSCompletionService.getInstance(element.project)